proper bill adding (#8)

+++ app/src/main/java/com/example/splitbill/ui/Screens/AddBillScreen/AddBillUiState.kt
@@ -6,7 +6,7 @@ import com.example.splitbill.data.local.entity.Friend
 data class AddBillUiState(
     val title: String = "",
     val friends: List<Friend> = emptyList(),
-    val selectedFriendIds: Set<Int> = emptySet(),
+    val selectedFriendIds: Set<Long> = emptySet(),
     val isLoading: Boolean = false,
     val error: String? = null,
     val saveSuccess: Boolean = false
